+++ tests/sim/status_warehouse/container/column/test_get_methods_column.py
@@ -9,6 +9,7 @@ class TestGetMethodsColumn(TestCase):
     def setUp(self):
         self.warehouse = Warehouse()
         self.column_config = ColumnConfiguration(
+            length=200,
             height = 325,
             x_offset = 125,
             width = 250,