+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Added a new exception in `load` if the offset is not found.
 - Added default config path if none is specified (`configuration/sample_config.yaml`).
 - Added total simulation time.
+- Added `length` field in the configuration.
 - Changed the first trayEntry reference. Now the first trayEntry will be the lower limit.
   For example, a tray with three entries; in the container the first trayEntry will be position 2, 
   and entries 0 and 1 are simple trayEntries.